+++ pages/profile/index.tsx
@@ -8,7 +8,7 @@ import { NextPageWithLayout } from "types/common";
 
 const ProfilePage: NextPageWithLayout = () => {
   /* 벡엔드연동 시 user 정보를 받아옴 */
-  const userInfo = true;
+  const userInfo = false;
 
   useEffect(() => {
     //user 정보가 없으면 로그인 페이지로 이동
